chore: rename activations -> glu_activations
